requirements.txt
absl-py==0.12.0
astor==0.8.1
cached-property==1.5.2
dataclasses==0.8
gast==0.4.0
google-pasta==0.2.0
grpcio==1.38.0
h5py==2.10.0
importlib-metadata==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
opencv-python==4.5.2.54
protobuf==3.17.2
PyYAML==5.4.1
scipy==1.5.4
six==1.16.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
